yx86/watchdog/hardlockup/hpet: Introduce a target_cpumask
The HPET hardlockup detector uses struct hpet_hld_data.monitored_cpumask
to keep track of the CPUs that must be monitored for hard lockups. When
the HPET timer expires, the CPU handling the interrupt issues an IPI to
a subset of the monitored CPUs. For instance, the IPI will be sent only
to the monitored CPUs in the same package. Let us call this subset of
monitored CPUs the target CPUs. Furthermore, in order to monitor all the
required CPUs, it is necessary to update the target CPUs in a round-robin
manner.

Start new IRQ work to 1) update the CPU handling the HPET interrupt to the
first CPU in the next package with a monitored CPUs; and 2) update the
target CPUs so it includes all the monitored CPUs in such package.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>