watchdog/hardlockup/hpet: Group packages receiving IPIs when needed
In order to keep the HPET interrupts of the hardlockup detector at a rate
of one per second or less frequent, the HPET timer only targets a single
CPU, the handling CPU. Then, such CPU issues an IPI to a subset of the
CPUs that the hardlockup detector monitors. This subset is composed of all
the monitored CPUs that are in the same package as the handling CPU.

However, there may be situations in which it is not possible to keep the
aforementioned HPET interrupt rate. This may happen if, for instance,
watchdog_thresh is set to 1 second and there are more than one socket in
the system. In such case, the HPET timer should expire 1/nr_sockets
seconds.

It is possible to keep the HPET timer expiration at one second or less
frequent if the packages receiving the IPI are grouped together. Hence,
if, for instance, watchdog_thresh is one second and there are two sockets,
monitored CPUs in the two sockets would receive the IPI interrupt at the
same time.

This approach has the drawback of increasing the communication between
sockets. However, cases in which we should group together packages should
be rare: only when watchdog_thresh is less than the number of sockets in
the system.

Implement functionality to use the logic above: when the hardlockup
detector is enabled in a CPU, check if grouping is necessary based in the
value of watchdog_thresh. Also update the timer period if needed. Lastly,
when updating target_cpumask, do it as many times as packages in the
group.

Signed-off-by: Ricardo Neri <ricardo.neri-calderon@linux.intel.com>